Make eval_model() work for multi-output
Also:
- Add missing parameters to config files.
- Move make_weights_function to utils.py
