+++ parameters/delphes-gnn-skipconn-onecycle.yaml
@@ -43,11 +43,19 @@ setup:
   num_epochs: 250
   num_val_files: -1
   dtype: float32
-  sample_weights: none
   trainable: all
   multi_output: yes
   classification_loss_type: categorical_cross_entropy
   lr_schedule: onecycle  # exponentialdecay, onecycle
+
+sample_weights:
+  cls: none
+  charge: none
+  pt: none
+  eta: none
+  sin_phi: none
+  cos_phi: none
+  energy: none
 
 parameters:
   model: gnn
