+++ parameters/delphes-gnn-skipconn.yaml
@@ -37,11 +37,19 @@ setup:
   num_epochs: 400
   num_val_files: -1
   dtype: float32
-  sample_weights: none
   trainable: all
   multi_output: no
   classification_loss_type: categorical_cross_entropy
   lr_schedule: exponentialdecay  # exponentialdecay, onecycle
+
+sample_weights:
+  cls: none
+  charge: none
+  pt: none
+  eta: none
+  sin_phi: none
+  cos_phi: none
+  energy: none
 
 parameters:
   model: gnn
